corrijo precio de int a double

+++ Models/Producto.cs
@@ -9,19 +9,19 @@ namespace tl2_tp6_2024_Isas321.Models
     {
         private int idProducto;
         private string descripcion;
-        private int precio;
+        private double precio;
 
         // Constructor vacío
         public Producto() { }
 
         // Constructor sin ID la base de datos lo autoincrementará
-        public Producto(string descripcion, int precio)
+        public Producto(string descripcion, double precio)
         {
             this.descripcion = descripcion;
             this.precio = precio;
         }
 
-        public Producto(int idProducto, string descripcion, int precio)
+        public Producto(int idProducto, string descripcion, double precio)
         {
             this.idProducto = idProducto;
             this.descripcion = descripcion;
@@ -31,7 +31,7 @@ namespace tl2_tp6_2024_Isas321.Models
         // Propiedades
         public int IdProducto { get => idProducto; set => idProducto = value; }
         public string Descripcion { get => descripcion; set => descripcion = value; }
-        public int Precio { get => precio; set => precio = value; }
+        public double Precio { get => precio; set => precio = value; }
 
         // Método para mostrar detalles del producto
         public string MostrarProducto()
